requirements.txt
RPi.bme280
numpy
psycopg2-binary
pyserial
smbus2
ruff
